+++ tests/runtime/grpc_server/conftest.py
@@ -22,15 +22,11 @@ import alog
 from caikit.core import MODEL_MANAGER
 from caikit.core.data_model.dataobject import render_dataobject_protos
 from caikit.runtime.grpc_server.grpc_server import RuntimeGRPCServer
-from caikit.runtime.grpc_server.servicers.global_predict_servicer import (
-    GlobalPredictServicer,
-)
-from caikit.runtime.grpc_server.servicers.global_train_servicer import (
-    GlobalTrainServicer,
-)
 from caikit.runtime.model_management.loaded_model import LoadedModel
 from caikit.runtime.service_factory import ServicePackage, ServicePackageFactory
 from caikit.runtime.service_generation.rpcs import TaskPredictRPC
+from caikit.runtime.servicers.global_predict_servicer import GlobalPredictServicer
+from caikit.runtime.servicers.global_train_servicer import GlobalTrainServicer
 from tests.conftest import temp_config
 
 log = alog.use_channel("TEST-GRPCCONF")
@@ -58,23 +54,6 @@ def runtime_grpc_test_server(open_port, *args, **kwargs):
                 # Give tests access to the workdir
                 server.workdir = workdir
                 yield server
-
-
-@pytest.fixture(scope="session")
-def sample_predict_servicer(sample_inference_service) -> GlobalPredictServicer:
-    servicer = GlobalPredictServicer(inference_service=sample_inference_service)
-    yield servicer
-    # Make sure to not leave the rpc_meter hanging
-    # (It does try to clean itself up on destruction, but just to be sure)
-    rpc_meter = getattr(servicer, "rpc_meter", None)
-    if rpc_meter:
-        rpc_meter.end_writer_thread()
-
-
-@pytest.fixture(scope="session")
-def sample_train_servicer(sample_train_service) -> GlobalTrainServicer:
-    servicer = GlobalTrainServicer(training_service=sample_train_service)
-    yield servicer
 
 
 @pytest.fixture(scope="session")
@@ -114,18 +93,6 @@ def training_management_stub(runtime_grpc_server) -> Type:
         runtime_grpc_server.make_local_channel()
     )
     return training_management_stub
-
-
-@pytest.fixture
-def sample_task_unary_rpc(sample_inference_service: ServicePackage) -> TaskPredictRPC:
-    return sample_inference_service.caikit_rpcs["SampleTaskPredict"]
-
-
-@pytest.fixture
-def sample_task_streaming_rpc(
-    sample_inference_service: ServicePackage,
-) -> TaskPredictRPC:
-    return sample_inference_service.caikit_rpcs["ServerStreamingSampleTaskPredict"]
 
 
 def register_trained_model(
